chore(docker): update Dockerfile and docker-compose for streamlined build context
- Adjusted Dockerfile to copy entire root context instead of only portal directory.
- Modified docker-compose to use updated Dockerfile path and root `.env` file.
- Simplified volume mapping for portal public assets.
