+++ apps/portal/src/utils/images.ts
@@ -4,3 +4,7 @@ export function getImageUrl(
 ) {
 	return `${baseUrl}/api/documents/${documentId}/data`;
 }
+
+export function getIconUrl(iconId: string, { baseUrl }: { baseUrl: string }) {
+	return `${baseUrl}/api/icons/${iconId}/icon`;
+}
